move unit test resource to pipeline folder
